# docs/swagger.yaml
definitions:
  Address:
    properties:
      address:
        type: string
      balance:
        type: string
      flipsCount:
        type: integer
      reportedFlipsCount:
        type: integer
      stake:
        type: string
      txCount:
        type: integer
    type: object
...
  Error:
    properties:
      message:
        type: string
    type: object
...
  Response:
    properties:
      error:
        $ref: '#/definitions/Error'
        type: object
      result:
        type: object
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
paths:
...
  /Address/{address}:
    get:
      operationId: Address
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                result:
                  $ref: '#/definitions/Address'
              type: object
        "400":
          description: Bad request
        "429":
          description: Request number limit exceeded
        "500":
          description: Internal server error
        "503":
          description: Service unavailable
      tags:
      - Address
...
swagger: "2.0"